+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Throw parser error when .json is accessed on response with invalid body.
 - Set `Content-Type: application/json` for `:http_request(..., {json = ...})` requests.
 
 # 0.5.1